research/arxiv_papers/MindFlow/requirements.txt
os
platform
mindspore.dataset
numpy
pandas
sklearn.preprocessing 
gc
psutil
mindspore.nn
sklearn.model_selection
model
mindspore.ops
mindspore